feat!: complete rework of routes emitter
The routes emitter has been reworked to be more concise, usable and practical. Adding the server root-URL to `getUrl` responses has been removed to allow for more flexibility. Auth information is now more complete and granular. Typeguards-in-routes has been removed for the time being, as it was clunky, didn't work very well overall and failed to achieve it's intended goal.

BREAKING CHANGE: This breaks existing configurations and - depending on how it was used - usage of the emitted routes object.